[development/file_reorganisation] moving cross platform path operations out of win_path and into a file of its own
